refactor(cli): extract badge parser owner
